requirements.txt
pycryptodome==3.19.0
PySocks==1.7.1
requests[socks]==2.31.0